+++ app/src/main/java/com/danilkinkin/buckwheat/editor/Editor.kt
@@ -47,9 +47,7 @@ fun Editor(
                 ) { focusController.focus() }
         ) {
             EditorToolbar()
-            if (mode == EditMode.EDIT) {
-                DateTimeEditPill()
-            }
+            DateTimeEditPill()
             CurrentSpendEditor(
                 modifier = Modifier
                     .fillMaxWidth()